refactor: Moved error output formatter
