+++ lib/src/better_command_runner/better_command.dart
@@ -4,6 +4,7 @@ import 'dart:io' show Platform;
 import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:cli_tools/config.dart';
+import 'package:cli_tools/src/config/output_formatting.dart';
 
 import 'better_command_runner.dart';
